+++ jtransc-main/test/big/AsyncIOTest.kt
@@ -1,0 +1,36 @@
+package big
+
+import com.jtransc.JTranscSystem
+import com.jtransc.JTranscSystemProperties
+import com.jtransc.io.async.JTranscAsyncFile
+import com.jtransc.io.async.readAll
+import com.jtransc.io.async.writeBytes
+import kotlin.coroutines.experimental.Continuation
+import kotlin.coroutines.experimental.CoroutineContext
+import kotlin.coroutines.experimental.EmptyCoroutineContext
+import kotlin.coroutines.experimental.startCoroutine
+
+class AsyncIOTest {
+	companion object {
+		private fun mainAsync(callback: suspend () -> Unit) {
+			var completed = false
+			callback.startCoroutine(object : Continuation<Unit> {
+				override val context: CoroutineContext = EmptyCoroutineContext
+				override fun resume(value: Unit) {
+					completed = true
+				}
+
+				override fun resumeWithException(exception: Throwable) = exception.printStackTrace()
+			})
+			if (!JTranscSystem.hasEventLoop()) {
+				while (!completed) Thread.sleep(10L)
+			}
+		}
+
+		@JvmStatic fun main(args: Array<String>) = mainAsync {
+			val tmpfile = JTranscSystemProperties.tmpdir() + "/AsyncIOTest.bin"
+			JTranscAsyncFile(tmpfile).writeBytes(byteArrayOf(1, 2, 3, 4))
+			println(JTranscAsyncFile(tmpfile).readAll().toList())
+		}
+	}
+}