+++ jtransc-rt-core/src/com/jtransc/io/JTranscAsyncIO.java
@@ -1,5 +1,0 @@
-package com.jtransc.io;
-
-public class JTranscAsyncIO {
-	// @TODO: Add this!
-}