+++ jtransc-rt-core/src/com/jtransc/io/JTranscFileStat.java
@@ -1,0 +1,7 @@
+package com.jtransc.io;
+
+public class JTranscFileStat {
+	public String name;
+	public boolean exists;
+	public long length;
+}