+++ jtransc-rt-core/src/com/jtransc/io/JTranscSyncIO.java
@@ -1,12 +1,15 @@
 package com.jtransc.io;
 
+import com.jtransc.JTranscSystem;
 import com.jtransc.annotation.*;
 import com.jtransc.annotation.haxe.HaxeAddMembers;
 import com.jtransc.annotation.haxe.HaxeMethodBody;
 import com.jtransc.annotation.haxe.HaxeMethodBodyList;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.RandomAccessFile;
 
 @JTranscAddFile(target = "js", priority = 0, process = true, prepend = "js/io.js")
 public class JTranscSyncIO {
@@ -53,7 +56,8 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return N.lnew(Math.max(0, FileStat.statSync(N.istr(p0)).size));"),
 		})
 		public long getLength(String file) {
-			return 0L;
+			checkNotJTransc();
+			return new File(file).length();
 		}
 
 		@Override
@@ -64,7 +68,8 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "new File(N.istr(p0)).deleteSync();"),
 		})
 		public boolean delete(String file) {
-			return false;
+			checkNotJTransc();
+			return new File(file).delete();
 		}
 
 		@Override
@@ -120,7 +125,13 @@ public class JTranscSyncIO {
 			}),
 		})
 		public int getBooleanAttributes(String file) {
-			return 0;
+			checkNotJTransc();
+			File f = new File(file);
+			int res = 0;
+			if (f.exists()) res |= 1;
+			if (f.isDirectory()) res |= 2;
+			if (f.isFile()) res |= 4;
+			return res;
 		}
 
 		@Override
@@ -149,6 +160,12 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return new File(N.istr(p0)).existsSync() || new Directory(N.istr(p0)).existsSync();"),
 		})
 		public boolean checkAccess(String file, int access) {
+			checkNotJTransc();
+			boolean result = true;
+			File f = new File(file);
+			result &= (access & 4) == 0 || f.canRead();
+			result &= (access & 2) == 0 || f.canWrite();
+			result &= (access & 1) == 0 || f.canExecute();
 			return false;
 		}
 
@@ -172,7 +189,8 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "try { new Directory(N.istr(p0)).createSync(); return true; } catch (e) { return false; }"),
 		})
 		public boolean createDirectory(String file) {
-			return false;
+			checkNotJTransc();
+			return new File(file).mkdir();
 		}
 
 		@Override
@@ -197,7 +215,8 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "new File(N.istr(p0)).renameSync(N.istr(p1));")
 		})
 		public boolean rename(String fileOld, String fileNew) {
-			return false;
+			checkNotJTransc();
+			return new File(fileOld).renameTo(new File(fileNew));
 		}
 
 		@Override
@@ -238,7 +257,8 @@ public class JTranscSyncIO {
 			}),
 		})
 		public String[] list(String file) {
-			return new String[] {};
+			checkNotJTransc();
+			return new File(file).list();
 		}
 
 		@Override
@@ -254,6 +274,7 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return N.str(Directory.current.path);"),
 		})
 		public String getCwd() {
+			//return new File(".").getAbsolutePath();
 			return cwd;
 		}
 
@@ -291,6 +312,7 @@ public class JTranscSyncIO {
 	})
 	static private class JTranscIOSyncFile extends ImplStream {
 		int mode;
+		private RandomAccessFile raf;
 
 		public JTranscIOSyncFile() {
 			init();
@@ -405,7 +427,15 @@ public class JTranscSyncIO {
 			}),
 		})
 		private boolean _open(String name, int mode) {
-			return false;
+			checkNotJTransc();
+			boolean readOnly = (mode & 2) == 0;
+			String smode = readOnly ? "r" : "rw";
+			try {
+				this.raf = new RandomAccessFile(name, smode);
+				return true;
+			} catch (FileNotFoundException e) {
+				return false;
+			}
 		}
 
 		@HaxeMethodBody("_stream.syncioClose();")
@@ -417,6 +447,8 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "this.file.closeSync();"),
 		})
 		private void _close() throws IOException {
+			checkNotJTransc();
+			if (this.raf != null) this.raf.close();
 		}
 
 		@HaxeMethodBody("return _stream.syncioReadBytes(p0, p1, p2);")
@@ -428,7 +460,12 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return this.file.readIntoSync(p0.data, p1, p1 + p2);"),
 		})
 		private int _read(byte b[], int off, int len) {
-			return -1;
+			checkNotJTransc();
+			try {
+				return this.raf.read(b, off, len);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
 		}
 
 		@HaxeMethodBody("return _stream.syncioWriteBytes(p0, p1, p2);")
@@ -440,7 +477,13 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "this.file.writeFromSync(p0.data, p1, p1 + p2); return p2;"),
 		})
 		private int _write(byte b[], int off, int len) {
-			return -1;
+			checkNotJTransc();
+			try {
+				this.raf.write(b, off, len);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+			return len;
 		}
 
 		@HaxeMethodBody("return _stream.syncioPosition();")
@@ -452,7 +495,12 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return N.lnew(this.file.positionSync());"),
 		})
 		private long _getPosition() {
-			return 0L;
+			checkNotJTransc();
+			try {
+				return this.raf.getFilePointer();
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
 		}
 
 		@HaxeMethodBody("_stream.syncioSetPosition(p0);")
@@ -464,7 +512,12 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "this.file.setPositionSync(p0.toInt());"),
 		})
 		private void _setPosition(long pos) {
-
+			checkNotJTransc();
+			try {
+				this.raf.seek(pos);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
 		}
 
 		@HaxeMethodBody("return _stream.syncioLength();")
@@ -478,18 +531,28 @@ public class JTranscSyncIO {
 			@JTranscMethodBody(target = "dart", value = "return N.lnew(this.file.lengthSync());"),
 		})
 		private long _getLength() {
-			return 0L;
+			checkNotJTransc();
+			try {
+				return this.raf.length();
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
 		}
 
 		@HaxeMethodBody("_stream.syncioSetLength(p0);")
 		@JTranscMethodBodyList({
 			@JTranscMethodBody(target = "js", value = "this._stream.setLength(N.ltoFloat(p0));"),
 			//@JTranscMethodBody(target = "cpp", value = "if (this->file != NULL) ::ftruncate(fileno(this->file), p0);")
-			@JTranscMethodBody(target = "cs", value = "this.file.Length = p0;"),
+			@JTranscMethodBody(target = "cs", value = "this.file.SetLength(p0);"),
 			@JTranscMethodBody(target = "dart", value = "this.file.truncateSync(p0.toInt());"),
 		})
 		private void _setLength(long newLength) {
-			throw new RuntimeException("Not implemented");
+			checkNotJTransc();
+			try {
+				this.raf.setLength(newLength);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
 		}
 	}
 
@@ -710,5 +773,9 @@ public class JTranscSyncIO {
 			read(out, 0, count);
 			return out;
 		}
+	}
+
+	static void checkNotJTransc() {
+		if (JTranscSystem.isJTransc()) throw new RuntimeException("Operation not implemented in this target");
 	}
 }