+++ jtransc-rt-core/src/com/jtransc/io/async/JTranscAsyncHttpClient.java
@@ -1,0 +1,4 @@
+package com.jtransc.io.async;
+
+public class JTranscAsyncHttpClient {
+}