+++ jtransc-rt-core/src/com/jtransc/io/async/impl/JTranscAsyncFileSystemDefault.java
@@ -1,0 +1,6 @@
+package com.jtransc.io.async.impl;
+
+import com.jtransc.io.async.JTranscAsyncFileSystem;
+
+public class JTranscAsyncFileSystemDefault extends JTranscAsyncFileSystem {
+}